PICS-14089 added missing null check.

+++ pics-legacy/src/main/java/com/picsauditing/PICS/FeeService.java
@@ -247,10 +247,12 @@ public class FeeService {
         if (billingStatus.isUpgrade()) {
             ContractorTag ssBlockTag = getSSBlockContractorTag(contractor);
 
-            contractor.getOperatorTags().remove(ssBlockTag);
-            feeDAO.remove(ssBlockTag);
-            billingService.syncBalance(contractor);
-            calculateContractorInvoiceFees(contractor);
+            if (ssBlockTag != null) {
+                contractor.getOperatorTags().remove(ssBlockTag);
+                feeDAO.remove(ssBlockTag);
+                billingService.syncBalance(contractor);
+                calculateContractorInvoiceFees(contractor);
+            }
         }
     }
